docker/requirements.txt
opencv-python==4.11.0.86
numpy==1.19.5
mtcnn==0.1.1
tensorflow==2.5.0
Pillow==9.5.0
scikit-learn==1.0.2
fastapi==0.70.0
uvicorn[standard]==0.15.0
python-multipart
Pillow
requests
boto3